requirements.txt
line-bot-sdk
flask~=1.1.2
werkzeug>=0.15.4
python-dotenv~=0.14.0
pymysql~=0.10.0
sqlalchemy~=1.3.19
pandas~=1.1.0
requests~=2.24.0
flask_restful
flask_cors
flask_jwt_extended
flask_cache
redis~=2.10.6
pip~=20.2.2
wheel~=0.34.2
zipp~=3.1.0
certifi~=2020.6.20
chardet~=3.0.4
zlib~=1.2.11
six~=1.15.0
webencodings~=0.5.1
idna~=2.10
urllib3~=1.25.10
pyparsing~=2.4.7
setuptools~=49.2.1
numba~=0.51.2